Improved scaffolding
