+++ packages/express-did-auth/src/factories/logout-factory.ts
@@ -1,0 +1,10 @@
+import { SessionManager } from '../types'
+
+export default function logoutFactory(sessionManager: SessionManager) {
+  // this function assumes it is invoked after a middleware that injects the user did in the request object
+  return function(req, res) {
+    sessionManager.delete(req.user.did)
+
+    res.status(200).send()
+  }
+}
