+++ packages/express-did-auth/src/factories/request-auth-factory.ts
@@ -1,0 +1,11 @@
+import { ChallengeVerifier } from '../types'
+
+export default function requestAuthFactory (challengeVerifier: ChallengeVerifier) {
+  return function (req, res) {
+    const { did } = req.params
+
+    const challenge = challengeVerifier.get(did)
+
+    res.status(200).send({ challenge })
+  }
+}
